Completed Assignment 3 Task 7
'Fix the issue encountered using the “z-index”. Make sure that the
highest “z-index” on the entire page is the “z-index” of the badge
(hint: Maybe the Stacking Context is a helpful concept in this
situation, so you might have to add the “position” property again). '

The hint is misleading. Is the lesson not to trust the instructor and
challenge everything they say?